generate UPP for Characteristics display

+++ src/pseudo_hex.rs
@@ -1,12 +1,12 @@
 use std::collections::VecDeque;
 use std::string::String;
 
-pub fn num_to_pseudo_hex(num: usize) -> String {
+pub fn num_to_pseudo_hex(num: u8) -> String {
     let new_base = vec!['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E',
                         'F', 'G', 'H', 'J', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'U', 'V',
                         'W', 'X', 'Y', 'Z'];
 
-    let mut value = num;
+    let mut value = num as usize;
 
     if num == 0 {
         String::from("0")
@@ -32,7 +32,7 @@ fn test_num_to_pseudo_hex() {
     for idx in 0..base.len() {
         let mut str = String::new();
         str.push(base[idx]);
-        assert_eq!(num_to_pseudo_hex(idx), str);
+        assert_eq!(num_to_pseudo_hex(idx as u8), str);
     }
 
     assert_eq!(num_to_pseudo_hex(34), "10");
